Feat : Sidebar에 footer 추가 - Home의 footer 미디어쿼리 수정 - footer 컴포넌트에 color props 추가 Issue #525
